Move files into src. Move output into dist folder
